move helper to main file
