+++ __init__.py
@@ -1,6 +1,7 @@
+import re
 from mycroft import MycroftSkill, intent_file_handler, intent_handler
 from jikanpy import Jikan
-import helpers
+
 
 USERNAME_SETTING_KEY = 'username'
 
@@ -10,7 +11,7 @@ class Mal(MycroftSkill):
 
     def initialize(self):
         self.jikan = Jikan()
-        self.alias_helper = helpers.AnimeAliasHelper()
+        self.alias_helper = AnimeAliasHelper()
         self.mal_username = self.settings.get(USERNAME_SETTING_KEY)
 
     def on_settings_changed(self):
@@ -63,3 +64,18 @@ class Mal(MycroftSkill):
 def create_skill():
     return Mal()
 
+
+class AnimeAliasHelper():
+    def __init__(self):
+        super().__init__(self)
+        self.jikan = Jikan()
+    
+    def name_matches_anime(self, name, anime_id):
+        anime = self.jikan.anime(anime_id) # could cache this locally?
+        titles = [anime['title'], anime['title_english'], anime['title_japanese']]
+        titles.append(anime['title_synonyms'])
+        # strip punctuation and whitespace
+        titles_norm = [re.sub(r'([^\w]|_)', '', t.lower()) for t in titles]
+        name_norm = re.sub(r'([^\w]|_)', '', name.lower())
+        
+        return any(title.startswith(name_norm) for title in titles_norm)
